requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
jq==1.6.0
python-dotenv==1.0.0
requests==2.31.0
sqlean.py==0.21.8.4
urllib3==2.0.7
